tast-tests: Fix how shelf.LaunchedApps looks for shelf icons
Last step of shelf.LaunchedApps test is to find default app items in the
shelf, and it searches for a shelf icon node using button role and the
app name. This narrows down the node search params to also specify the
class name. Without this there is a chance that the test finds a node
that is not really in shelf - for example, AppListItemView nodes in
launcher would match search for a button with a name matching the app
name (and in tablet mode, launcher is in shown state).

BUG=None

TEST=tast run DUT shelf.LaunchedApps

Change-Id: Ia088d795e83f4373af3400d52d58cffc7a37cc84