autoupdate: Add basic N2N update test
The DUT image version does not change during this update.
This version is for manual execution. If it works well with lab execution,
it can be added as an informational test in a follow up CL.
For the testing I used a local setup of the infra Test Lab Wiring,
and a DUT leased from crosfleet. This way the DUT and the caching
server are at the same environment.

BUG=chromium:1240933
TEST=tast run -var=autoupdate.tlwAddress="<local_TLW>" <DUT> \
TEst=  autoupdate.BasicNToN

Change-Id: Ib7896e24cea8b8631e8c59ad72fb0cbe575fb2b2