webrtc: add some SVC RTCPeerConnectionPerf test cases
This CL adds a SW version of the VP9 k-SVC Perf test so that we
can compare with the existing HW-accelerated one (and also
verify the former as a bonus).

This CL also takes the chance to homogenise the test case names
to follow the Web API scalability modes names [1].

k-SVC can be forced via flags or used via the experimental Web
API: this CL adds a VP9 k-SVC Perf test via this experimental
API, which should show the same results as the existing one.

[1] https://www.w3.org/TR/webrtc-svc/#scalabilitymodes

BUG=b:194895782
TEST=tast run <volteer IP> webrtc.RTCPeerConnectionPerf.*l3t3*

Change-Id: Icb703629f617e84479352953b07d3177f0a1960b