tast-test: Close folder after rename in launcher.RenameFolder
RenameFolder function opens a folder in the app list, updates its name
using the folder UI textfield, and tries to find the folder using its
updated name in the apps grid.
This adds a step to close the folder (using esc accelerator) before
searching the folder item in the apps grid UI. Recently, apps grid view
logic was updated to hide items in the root apps grid view from
automation API (to prevent chromevox from traversing from an opened
folder UI to the root apps grid view), which broke tests that are using
launcher.RenameFolder (the last step of finding the updated folder item
in the root apps grid view was failing). After the folder gets closed,
the test should be able to find the folder item in the root apps grid
(and it's cleaner for a utility function to leave the apps container in
the same state as when the test starter - with no folders open).

BUG=chromium:1256925
TEST=tast run DUT launcher.*Folder*

Change-Id: I2b0e23f767091a06bcd92c879287bd2e6d5647cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3206856
Tested-by: Toni Barzic <tbarzic@chromium.org>
Reviewed-by: Matthew Mourgos <mmourgos@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Eliot Courtney <edcourtney@chromium.org>
Tast-Review: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Toni Barzic <tbarzic@chromium.org>