bios_service: Add BackupECRW() and RestoreECRW() methods to BiosService
This CL adds specialized calls to backup and restore EC_RW region of
EC firmware on DUT, to ensure safe writing over this region with an
additional restore calls if anything goes wrong during the test or
the testcase itself just fails.

This introduces additional ImageSection definition for EC_RW region
as well as protobuf structures for these new methods in BiosService.

BUG=b:194908258
TEST=tast run -var servo=$servoIP $dutIP firmware.ECHash

Change-Id: Ic7529b1a8819c673b0e528eb3f29a78417f2d349