crostini: Close VM log reader when closing the precondition
The log reader keeps open a file descriptor to the cryptohome, which
breaks unmounting the cryptohome in some configurations if it is kept
around. Fix this by tying its lifetime to the precondition, and
skipping over any log lines added before the precondition is opened.

Eventually this sort of this should be moved to a fixture, which makes
this sort of post-test processing easier.

BUG=b:202065620
TEST=CQ

Change-Id: I253c963df55344012c613f69813a11599512b23a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3212167
Commit-Queue: Fergus Dall <sidereal@google.com>
Tested-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Judy Wu <jinrongwu@google.com>