arcappcompat: launch in clamshell
Adjusted the fixture for clamshell tests to boot the device into
clamshell mode. This helps to ensure the device is in the correct state,
and eliminates issues when the device starts out in tablet mode, and a
pointer device is attempted to be used.

BUG=b:202442984
TEST=push apk && tast run ${HOST} arc.Standardized*

Change-Id: I4ce1ed5bffdf5b27bbe7e5690a372c1e2cb7d027
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3213548
Reviewed-by: Tetsui Ohkubo <tetsui@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: David Welling <davidwelling@google.com>
Commit-Queue: David Welling <davidwelling@google.com>