faft: Temporary logging to detect running in lab
Based on recent failures, and timeouts, we should not use ssh if we
can connect directly to servod. This logging should help determine the
right logic for detecting if we are running in skylab. In the lab, we
shouldn't need ssh to connect to the servo port. We do need it for
running tests locally against skylab duts.

BUG=b:190532175
TEST=None

Change-Id: I9709f782b229f5a68837fb506accd3394c422edd