wifi.APSupportedRates.11g: promote to wifi_matfunc
The test is not completely stable yet (95% pass rate) but
- same pass rate as the tauto version: http://shortn/_n8lU3OttpZ
- the vast majority of failures are not caused by the test itself
("Failed to collect pcap")

Given that, it makes sense to move the test to wifi_matfunc, get rid of
the tauto version, and focus on stabilizing the infrastructure. There's
no benefit in keeping the tauto version in wifi_matfunc and the tast
version in the "unstable" suite.

BUG=b:160281713
TEST=./platform/tast/fast_build.sh -T
TEST=tast list $DUT '("group:wificell" && wificell_func &&
!wificell_unstable)' | grep APSupport
wifi.APSupportedRates.11g
TEST=tast list $DUT '("group:wificell" && wificell_func &&
wificell_unstable)' | grep APSupport
wifi.APSupportedRates.11ac

Change-Id: I3abde50979255430b774b5ec646632e8dc8a7f22