wifi.SimpleConnect.8021xwpa3: promote to wifi_matfunc
The test has been stable for a while. When it fails it's either an
infrastructure problem (AP is not reachable over SSH) or the DUT has
previously been broken by another test (b/202565544) in the "unstable"
pool.

Since the test itself is not at fault, promote it to wifi_matfunc.

BUG=b:186590933
TEST=~/trunk/src/platform/tast/tools/go.sh test -count=1
chromiumos/tast/remote/bundles/cros/wifi

Change-Id: I3aec1184402d09f9125c3be4b56b8fe00bb8ca02