faft: Detect serial AP firmware, and adjust sleep
If the AP firmware has the serial port enabled, which is common for
bringup, but not common otherwise, increase the boot time delays. This
should help with getting FAFT to pass on new machines, and avoid people
needing to change the firmware_screen delay in the json files to really
large values for new models.

BUG=b:202841102
TEST=Ran firmware.Fixture* on hayato w/ serial AP

Cq-Depend: chromium:3217864
Change-Id: I3bd20b27328fc8134b1852aa377b553b688dfd86
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3219410
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Andrew Luo <aluo@chromium.org>