Fix audio.UIInput.gain when external audio device is present
When an external audio device is present (i.e. speaker or headphones
plugged in), the list of audio options in Quick Settings overflows. To
click one that is off-screen, you need to scroll to it, and an effective
way to do this is with the FocusAndWait function. However, the a11y tree
reports incorrect locations for the previously-overflowed nodes, which
causes mouse clicks to be sent to the wrong locations.

In audio.UIInput.gain, this caused the wrong audio option (Mic jack) to
be clicked instead of the desired "Loopback Capture" option. To fix
this, we can just use DoDefault on the desired audio option, as it does
not depend on screen location. This test is more about the integration
between audio and the UI, so verifying the correctness of UI locations
is not particularly relevant.

BUG=b:197044199
TEST=tast run <dut> audio.UIInput.gain (with headphones plugged in to
DUT)

Change-Id: I34f11247aeb542a8ab80ad3d1a22ccb65c0dc94b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3219721
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Auto-Submit: Kyle Shimabukuro <kyleshima@chromium.org>
Tast-Review: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>