faft: Don't retry WaitConnect, it retries itself
In stainless I observed a test that failed, where it attempted to
connect via ssh every 5 seconds, and as soon as the poll loop failed,
the cleanup code connected just fine.

https://stainless.corp.google.com/logview/chromeos-test-logs/test-runner/prod/2021-10-12/c8c54431-6295-4cae-aaec-53d00dc82f91

Don't retry WaitConnect, since it has internal retries. Also 3s might
not be long be enough time for what it is trying to do.

BUG=b:174800291
TEST=tast run $d firmware.Eventlog.normal

Change-Id: I18fddcd8ecc55eedce33582ddef33ee012c070bf