tast-tests: Fail DataCollector and CacheValidation on non-user build of ARC
The packages installed in userdebug and eng build of ARC do not have
the expected signature and cache validation always fails. In this
change we're adding a fatal error to terminate the test early when it
is run against a userdebug or eng build.

BUG=b:202845451
TEST=tast run DUT arc.CacheValidation.r fails on userdebug and passes
on user build
TEST=tast run DUT arc.DataCollector.vm_local fails on userdebug and
passes on user build

Change-Id: I856e7e8217e7bc5a85835f8b91aa375c564fd291
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3219722
Tested-by: Muhammad Hasan Khan <mhasank@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Yury Khmel <khmel@chromium.org>
Commit-Queue: Muhammad Hasan Khan <mhasank@chromium.org>