nearbyshare: fix UI finder for the "ACCEPT" button
In-contacts tests are failing on M96 (and soon M97) due to an unrelated
change to internal UI details. The tests are failing to find the
"ACCEPT" button on the incoming share notification, since we were trying
to find it by a combination of name + classname. This is the second time
we've had to update it for this reason, so I've changed the finder to
remove use of internal details, and instead use things that come from
the a11y tree like the Name and Role attributes. Hopefully this will be
more stable in the long-term.

Also removed the name regex that was being used to account for version
skew, since this is no longer an issue in the current multi-DUT lab
environment.

BUG=b:198166431
BUG=b:187795854
TEST=tast run <dut> nearbyshare.PhoneToCrosInContacts*

Change-Id: Ibe6acd8f66f265a1147831cfba8e2bee08ad6897
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3219723
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Auto-Submit: Kyle Shimabukuro <kyleshima@chromium.org>
Tast-Review: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>
(cherry picked from commit 270dd59b4cfed0531a099a322ede311e40cca668)