printpreview: Replace nil Action
Because the tast test and the Android print test app are not
synchronized, WaitForPrintPreview needs to deal with both cases where
loading text appears slowly or appears and disappears very quickly.  It
wraps a WaitUntilExists call in IfSuccessThen to try to ignore errors,
but it passes a nil Action.  When the inner call occasionally succeeds,
this results in attempting to treat nil as a function pointer.  Rather
than passing nil to do nothing, pass in a real Action that does a very
short Sleep.

BUG=b:202836018
TEST=15 runs each of `tast run arc.Print.vm` and `tast run arc.Print`

Change-Id: Iff2398c30ff96e973f3c19df1533bce687f0c194
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3218278
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Pranav Batra <batrapranav@chromium.org>
Reviewed-by: David Welling <davidwelling@google.com>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>