video/data/tools: Add .external file tools
Tast uses .external files to store a reference to files in Google
Storage. The files have a URL, size and checksum. The URL typically
contains a  timestamp.

Add a script to generate or update a *.external file from a local file.
Also, add a script to validate *.external files by downloading the
remote file and validating the size and checksum.

BUG=None
TEST=Used tools to upload H.264 test vectors for b:202968780

Change-Id: I6a2c9e8e678d668c6e43860a29ff038befb00ade