video.Contents: Poll for fullscreen status.
This CL makes the video.Contents tests poll for fullscreen status and
retry requesting the video to go fullscreen as needed. Before this CL,
the tests would inject the 'f' keyboard input in order to trigger a JS
event handler that requests the video to go fullscreen, and this would
happen only once. The problem with this is that Chrome is not
necessarily restarted between tests, so in the transition from one test
to the next, inputs may get ignored for a little bit. With this CL, we
(1) inject the 'f' key; (2) check if the video went fullscreen; (3)
retry if not (every second for a maximum of 5 seconds).

BUG=b:199393037
TEST=video.Contents.{h264_1080p_hw,h264_360p_exotic_crop_hw} on hatch

Change-Id: I27b94eb14a58652e5416298225128d7ef0ee755b