wifi: check the "enabled" field of Intel SAR tables
WRDS and EWRD tables contain an "enabled" field that controls whether
the table will be used (if set to 1) or ignored in favor of OTP values
or other configured SAR tables.
Only check the validity of the raw values in cases where the "enabled"
filed is set to 1.

BUG=None
TEST=tast run ${DUT} wifi.CheckIntelSARTable
TEST=CQ and out-of-band wifi_matfunc to test across the fleet.

Change-Id: Ia3f6abb3065f23dfb150bcfb268bf148ba8161f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3228894
Tested-by: Nicolas Norvez <norvez@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Kevin Lund <kglund@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>