tast-tests: Adds initial Lacros Primary tests
We are heading over to Lacros Primary soon but there are
still users using Lacros side-by-side.
This change will add a small number of Lacros Primary specific
tests and a new fixture that enables Lacros as a primary browser.

If this works okay we can gradually roll this out for other Tast
tests to use it too targeting M97.

This also removes rootfs specific tests that have been already
migrated.

BUG=chromium:1260037
TEST=tast run lacros.ShelfLaunch.primary
TEST=tast run lacros.ShelfLaunch*
TEST=tast run a11y.Smoke.lacros_primary
TEST=tast run a11y.Smoke*

Change-Id: I1e84e8c58b921f0a3c912e3b7af9883f56f68bf4