tast-tests: Adds initial Lacros Primary tests
We are heading over to Lacros Primary but there are
still users using Lacros side-by-side.
This change will add a subset of Lacros Primary specific
tests and a new fixture that enables Lacros as a primary browser.

If it works okay we can gradually roll this out for other Tast
tests to use it too.

This also includes minor fixes:
- Add a workaround to ShelfLaunch to reduce a flake in VM tests
that is caused by a new browser window popped up on startup if lacros
is a primary browser.
- Remove rootfs fixture and specific tests that have
been already migrated.

BUG=chromium:1260037
TEST=tast run lacros.ShelfLaunch.primary
TEST=tast run lacros.ShelfLaunch*
TEST=tast run a11y.Smoke.lacros_primary
TEST=tast run a11y.Smoke*
TEST=tast run '("group:mainline" && !informational && "dep:lacros")'

Change-Id: I1e84e8c58b921f0a3c912e3b7af9883f56f68bf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3228591
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Hyung Tae Kim <hyungtaekim@chromium.org>
Commit-Queue: Hyung Tae Kim <hyungtaekim@chromium.org>