video: Use re-timestamped videos for video.PlaybackPerf.*
video.PlaybackPerf.* is supposed to use 30 or 60 fps videos, however
many sequences are actually slightly less than that (e.g. 29.97fps
ISO 30). This CL corrects that by pointing to new sequences that I
retimestamped ("conformed" in the lingo) to be exactly one of those
frame rates.

For reference the command used for conforming is:
  ffmpeg -v error -i input_file -filter:v fps=30 output_file

which basically reencodes the video (adjusting the audio too, if
needed), see [1, 2] (and the explanation linked from [2]).

[1] http://underpop.online.fr/f/ffmpeg/help/fps.htm.gz
[2] https://trac.ffmpeg.org/wiki/ChangingFrameRate

(There was other scripting involved in fixing every test, and
I used tast-tests/tools/generate_external_file.py to help creating
the .external files in batch).

BUG=b:202853614
TEST=tast run <volteer> video.PlaybackPerf.* still works

Change-Id: Iee3bcc32c31d9db8dd8c13e212b7f0fad41161a6