Setting ui.QuickCheckCUJ2 to be picked up by crosbolt nightly.
The Performance CUJs are required to be run on the Crosbolt.
This CL is setting the parameters for QuickCheckCUJ2 to be picked up by
Crosbolt nightly. The frequency will be changed to fortnightly in the
following CL, but it is set to nightly temporarily to observe the
effects

BUG=b:188604416
TEST=None

Change-Id: Ibbff2c9111f8df2aa71c181d83d044828c8b5ad4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3229330
Tested-by: Afshin Sadrieh <asadrieh@google.com>
Auto-Submit: Afshin Sadrieh <asadrieh@google.com>
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Reviewed-by: Alex Bergman <abergman@google.com>
Commit-Queue: Alex Bergman <abergman@google.com>