network: DNSProxy: Wait for the settings propagation properly
DNS queries through the proxy sometimes are successful even if
the network is blocked. This is because the settings or the
network configuration that is not yet updated. Previous patch
tests this behavior by adding a timeout: crrev.com/c/3184845

This patch removes the arbitrary delay added by waiting for
shill property and iptables output.

This patch also adds a connectivity check to distinguish
between DNS failures and connectivity failures.

BUG=b:199093160
TEST=tast run ${DUT} network.DNSProxy.*
TEST=tast run ${DUT} network.DNSProxyOverVPN.*

Change-Id: Ibf8bd775fd2b25a7c3ee1656bb911de112e59150