network: DNSProxy: Wait for the settings propagation properly
DNS queries through the proxy sometimes are successful even if
the network is blocked. This might be caused because the
settings or the network configuration is not updated yet.

This patch replaces the arbitrary delay added by
crrev.com/c/3184845 to delay by waiting for shill property and
iptables output.

This patch also adds a connectivity check to distinguish
between DNS failures and connectivity failures.

BUG=b:199093160
TEST=tast run ${DUT} network.DNSProxy.*
TEST=tast run ${DUT} network.DNSProxyOverVPN.*

Change-Id: Ibf8bd775fd2b25a7c3ee1656bb911de112e59150