faft: Add test to verify servo connectivity
In firmware.DeepSleep, the ssh connection sometimes dies during the
test. This test attempts to prove that you don't need ssh forwarding to
connect to servo when you are running in the lab. Unfortunately we need
it to be submitted in order to run it in the lab.

BUG=b:190532175
TEST=tast run --var servo localhost:9999 $d firmware.ServoNetwork

Change-Id: Ic2e264bf50db9cb26558222ef07eb24e38f547ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3232161
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Kshitij Shah <tij@google.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Auto-Submit: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Kshitij Shah <tij@google.com>