faft: Allow "Wake" events in recovery->normal boot
Since most platforms are failing on tast.firmware.Eventlog.rec_normal
with the error "Incorrect event logged" and the event is either
"ACPI Wake | S5" or "ACPI Wake | Deep S5" that must be normal. That is
the event that is allowed by the lenoa specific version, so we don't
that variant any more.

This test case didn't exist in the Tauto version of this test, so it's
not clear what the correct behavior is.

BUG=b:174800291
TEST=tast run $d firmware.Eventlog*

Change-Id: I8813cd5b0f60e8078087c83ab517be8c98f03d88