roam_ft: remove DiscoverBSSID
When this test used to use the Roam D-Bus command to force a roam, this
step was needed to detect the target AP before roaming. BSS TM Request
will request a scan on its own if the target AP is not in its cached
scan results, so we no longer need this extra step.

BUG=b:188468135
TEST=`tast run ${DUT} wifi.RoamFT.*`

Change-Id: I1b27d5229732d9de7c1e612f89a6ba0edab45be1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3232167
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Kevin Lund <kglund@google.com>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>