Reorder disconnect in roam tests
Deferrals are executed LIFO. The ordering of these roam tests is as
follows:

1. Configure AP1 (and defer teardown of AP1)
2. Connect to AP1 (and defer disconnection)
3. Configure AP2 (and defer teardown of AP2)
4. Roam from AP1 to AP2
5. End test

Since we end up connected to AP2 and we tear AP2 down before
disconnecting it, we get deauthed and the subsequent deferred disconnect
fails. Move the deferred disconnection after the deferred teardown of
AP2 so that we disconnect first.

BUG=b:203419921
TEST=run all modified tests

Change-Id: Ia12b15a46fdb565c57d661c5a5321512b7fe11ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3232168
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Matt Stark <msta@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Arowa Suliman <arowa@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>