faft: When writing to usb stick, use nocreat opt
If the disk block device disappears between fdisk and dd, don't let dd
create a file. If you do, the code believes that dd finished writing to
the usb stick, but really it just took up a bunch of local disk space
and did nothing to the USB disk.

Also switch from oflag=dsync to conv=fsync.  I believe that will cause
the write to sync at the end of the write only, which might be faster.

BUG=b:169333650
TEST=Ran some dd cmds on DUT

Change-Id: I85973a41c41e284a6b36c4426c0bc7ab47371065