factory Kiosk: promote as critical test
As Kiosk test itself is running stable more than a month, promote it as
critical. The nocturne can not pass this test but since it is
not being manufactured, simply excluded this platform. We have some
failure but is unrelated to the test itself but the lab environment, we
suspect it could be model related, thus still make them informational.

BUG=b:203720382
TEST=tast run $(DUT) factory.Kiosk

Change-Id: I8743657d4a50cc9e15416111d16b898ffec52b90
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3233550
Commit-Queue: Sung-Chi Li <lschyi@chromium.org>
Tested-by: Sung-Chi Li <lschyi@chromium.org>
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Reviewed-by: Meng-Huan Yu <menghuan@chromium.org>
Reviewed-by: Stimim Chen <stimim@chromium.org>