arc: Insert short sleep in Downloads.vm to avoid adb push failure
arc.Downloads.vm has been consistently failing since virtio-fs was
enabled on the Downloads directory.
This is because the current configuration of virtio-fs has a cache sync
issue for a short period of time. As a result, adb push to the Android
path /storage/emulated/0/Download/capybara.jpg fails when it's executed
right after a file is removed from the corresponding Chrome OS path
/home/chronos/user/Download/capybara.jpg.
In order to avoid this adb push failure, insert a one-second sleep
between file removal and adb push.

BUG=b:182870415
TEST=tast run <DUT> arc.Downloads*

Change-Id: Icd130446ee607eb919855cff6d762437415f0566