faft: Don't use ssh to connect to servod ever
Connect ssh to servo host lazily.

Never use ssh port forwarding to connect to servod. The only reason to
use ssh forwarding is to connect to a skylab machine from your
workstation or cloudtop, but ssh from TAST to the lab doesn't work in
the chroot, so you have to redirect ports. If you are already
redirecting ports, why should this code?

Delete firmware.ServoNetwork test.

BUG=b:190532175
TEST=Ran firmware.ServoGBBFlags & firmware.DeepSleep on leased dut
and local dut

Change-Id: I1a231957e186886363c79ac55024ca36ab9a1448