Tast: Reintroduce support to cache extracted raw videos in encode tests
As the raw video files used to test video encoding are too large to
download and store on a DUT, videos are stored in the vp9.webm format.
Before each test these videos are extracted to create the raw input
video. As this process can take a lot of time crrev.com/c/2437751 was
created to add support to cache the extracted raw video files, which
saves about 30s test time when running 1080p video encode testcases.

Upon switching to the new VideoEncoder tests this features was lost for
some tests, so this CL reintroduces it. The feature is disabled by
default to avoid affecting disk space on lab devices, and can be
manually enabled to speed up local development.

Caching is reintroduced in the following Chrome video encode tests:
- video.EncodeAccel.*
- video.EncodeAccelPerf.*

BUG=b:169725788
TEST=various video encode tests on Hatch

Change-Id: I33efdee681c33d6ffd7c8008786799335b9e5ee9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3244549
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Judy Wu <jinrongwu@google.com>
Tested-by: David Staessens <dstaessens@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>