tast-tests: Create fixture that starts a Samba server
Introduce a fixture that starts up smbd and shares an ephemeral
directory that acts as a guestshare. The daemon is available throughout
the entire lifetime of the fixture and in between each test it will
clean the guestshare and unmount all the shares from Chrome.

BUG=chromium:1156844
TEST=With relation chain applied, tast run <DUT-IP> filemanager.SMB

Change-Id: Ifd06fde6bf8c37b1c86965ef8b05e9885eb5d7fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3244557
Tested-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Reviewed-by: David Stevens <stevensd@chromium.org>