tast-tests: Switch filemanager.SMB from Crostini to smbd
The new fixture smbStarted uses smbd to start a local Samba server. This
will allow the test to remove the flaky portion and should enable this
test to move to critical. Note: Most of the files being removed here are
related to the crostini dependency, no other tests in filemanager.*
namespace use a Crostini container.

BUG=chromium:1156844
TEST=With relation chain applied, tast run <DUT-IP> filemanager.SMB

Change-Id: I119a5d3b9aefcef86644bacc3d747fbd4868adb2