health: Check CPU model name
We didn't check CPU model name because it's optional. Besides, in the
past, we could only fetch CPU model name for Intel project. In
b:203752932, we find some approaches for ARM devices and report SoC
model name instead. So we want to check this field in the future.

BUG=b:203752932
TEST=tast run $dut2 health.ProbeCPUInfo

Change-Id: Ie2b98a55252efaa37164dd50abf50120047c65c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3245351
Tested-by: Kerker Yang <kerker@chromium.org>
Reviewed-by: Kamil Szarek <kamilszarek@google.com>
Commit-Queue: Kerker Yang <kerker@chromium.org>