Tast: Stop activity after running ARCVM tast video encode perf test.
After running the arc.VideoEncodeAccelPerf test the activity running
the test is not properly stopped and will keep on encoding, using up a
lot of CPU. When running another Tast test in isolation ARCVM is
restarted so the activity is automatically stopped. But when running
multiple tests in a single batch of tests this might influence other
test runs.

It's not immediately clear though whether this is (partly) responsible
for the issues we're seeing in b/190446278.

BUG=b:190446278
TEST=arc.VideoEncodeAccelPerf.h264_*_i420_vm on octopus

Change-Id: Ia0912e6858af7728731547a6f3f2a33429f570e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3245753
Tested-by: David Staessens <dstaessens@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Renuga Nanjappan <rnanjappan@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>