tast-tests: Update platform.Drivefs to use GAIALoginPool
The platform.Drivefs test is getting rate limited when using the
catapault login, so move it to using the login pool method instead. This
also takes the liberty to remove dats@ from the OWNERS (no longer on the
team) and add the team alias.

BUG=b:204268996
TEST=tast run <DUT-IP> platform.Drivefs

Change-Id: Ib1d4e8efaf5201ac72508273ada2d69816f3f5ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3246891
Tested-by: Ben Reich <benreich@chromium.org>
Reviewed-by: Austin Tankiang <austinct@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>