lockscreen.Signout: Click on the signout button repeatedly
Sometimes device might go to sleep while we wait for the CPU to become
idle. In that case the Sign out might ignore couple first clicks. So in
this CL we click on the button repeatedly. I reproduced the issue
locally with the sleeps in the tests - and the CL seems to fix this.

BUG=b:190654593
TEST=tast run $DUT lockscreen.Signout

Change-Id: I0841decff118312e0d7b7000867011c55d66a127