camera: Share fixture between CCAUI{DocumentScanning,QRCode,PTZ}
These tests actually require the identical setup but with different
scenes. Therefore, this CL tweaks the logic so that we can dynamically
switch the scene when using fake camera so that they can share the same
fixture and eventually reduce the overall test time.

This CL also:
1. Merge camera.CCAUIQRCode.* back to camera.CCAUIQRCode since we don't
need different fixtures now
2. Update all .y4m to use .mjpeg for fake camera stream due to smaller
size

BUG=b:204164385
TEST=tast run [DUT] camera.CCAUI*

Change-Id: Ie5ab649257e399c1f4f97e59140337789de504f2