Remove app name check from Gallery root finder.
Currently, the Gallery app is found by searching for its name in the
window title. In http://crrev/c/3244921 however, I'm adding the ability
for Gallery (MediaApp) to override this title and instead show the
current filename. This breaks tast tests which rely on the aforementioned
app title search.

To remedy that, this CL updates the way we find the Gallery root by
instead searching for a BrowserFrame; where there is only currently one.
To future proof against changes such as Files app becoming a SWA and
adding a BrowserFrame themselves, I have also added a check to see if
the Gallery name is a part of the BrowserFrame. This is consistent
in the a11y tree before and after http://crrev/c/3244921 lands.

BUG=b:200898614
BUG=b:202214335
TEST=Ran Gallery Tast tests

Change-Id: I4a83e6411a12b42dad111135b128e72a23c5e576