policy: fix UserAvatarImage test
Before this CL, the test is failing on almost all boards[1]. The test
is failing because the test logic depends on the existence of 'Cat'
avatar image, which is removed in newer cros versions.

This CL introduces the fix: retrieve the list of avatar images and
select one of them rather than depending on the 'Cat' one.

[1] https://cros-health-prod.googleplex.com/test/policy.UserAvatarImage

BUG=chromium:1185195, chromium:1257963
TEST=tast run <ip> policy.UserAvatarImage

Change-Id: Ice800ad3f4d72f84bf4c4e4a7c0c4d849bb42bdd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3247100
Tested-by: Mahmoud Gawad <mgawad@google.com>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Commit-Queue: Mahmoud Gawad <mgawad@google.com>