kiosk: Add support of Chrome Options
Kioskmode library accepts Chrome options and passes them to the Chrome
instance that starts in Kiosk mode. That way we remove the redundancy of
the flag for passing sign in key for extension on SingIn screen and
allow to pass extra arguments to Chrome e.g. to allow Lacros testing.

BUG=b:204525928
TEST=tast run <ip> kiosk.StartAppFromSignInScreen kiosk.FloatingAccessibilityMenuEnabled

Change-Id: I50ded5166eb003e468d07e187c1bb24113a7d877
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3252960
Tested-by: Kamil Szarek <kamilszarek@google.com>
Commit-Queue: Anqing Zhao <anqing@chromium.org>
Reviewed-by: Gabor Zsolt Magda <gabormagda@google.com>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>