kiosk: Add support of Chrome Options
Kioskmode library accepts Chrome options and passes them to the Chrome
instance that starts in Kiosk mode. That way we remove the redundancy of
the flag for passing sign in key for extension on SingIn screen and
allow to pass extra arguments to Chrome e.g. to allow Lacros testing.

BUG=b:204525928
TEST=tast run <ip> kiosk.StartAppFromSignInScreen kiosk.FloatingAccessibilityMenuEnabled

Change-Id: I50ded5166eb003e468d07e187c1bb24113a7d877