tast-tests: Higher interval to poll UI events
This is not a bugfix, but a preventive change to avoid
any flakes in polling UI events in Tast tests.

Currently WaitFor* functions in ash package check UI conditions
by executing JS codes every 100ms by default, which might be too
much for many tests that are not very time sensitive.
A higher interval may help reduce flake in polling.
Also, it adds a new WaitFor method with a certain interval of
polling time and a condition to be met.

BUG=chromium:1258664
TEST=fast_build.sh -T (unittest)
TEST=tast run ${DUT} '("group:mainline" && "dep:chrome" && !"dep:lacros" && !informational && !"dep:android*" && !"name:crostini.*")'
TEST=tast run lacros*

Change-Id: Ie720ed681c45c4b653fe23d7dc3a518ece5da67f