[TAST] Update class name used in Quick Settings tests.
This change updates the class name used for the Network feature pod
within the Quick Settings tests. This class has been renamed so that it
can be replaced as part of a revamp to make this code more testable.

Without: https://paste.googleplex.com/5109030704906240?raw
With:    https://paste.googleplex.com/6043475968524288?raw

BUG=chromium:1264078,chromium:1169479
TEST=manually

Cq-Depend: chromium:3255346
Change-Id: Ib9c0e4d45ccbea5c463cf963d88c0bcfedd949e1