Set onboarding complete in Nearby Share setup
Onboarding used to be set as complete as a side effect of
enabling Nearby Share. A recent change explicitly sets
onboarding complete on the front end when the onboarding
flow is complete, which broke the assumption in the
e2e test setup. This CL explicitly sets onboarding
complete in the test setup.

BUG=chromium:1265567
TEST=manually verified

Change-Id: Ieb3c58499471868567ce09fd5c1fe94df5a2a987