Update patch set 2
Patch Set 2:

From chrome_platform/tast/gerrit.gwsq:
Note: A shadow reviewer was assigned to this CL.
WARNING: gwsq was unable to find a reviewer who is not on vacation. As a fallback, gwsq is ignoring vacations and assigning mohamedaomar.
WARNING: gwsq was unable to find a reviewer who is not on vacation. As a fallback, gwsq is ignoring vacations and assigning mohamedaomar.
Shadowed: seewaifu

Reviewer source(s):
jbettis, mohamedaomar, seewaifu is from context

Patch-set: 2
CC: Gerrit User 1255833 <1255833@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>