Tast: Validate intel fw dump.
Make use of the fw-dump-extractor.pl that is provided by intel to check the
validity of the fw dump instead of checking the size directly. Only the
initial checkes in the are implemented:
1- Check the file size in the header.
2- Check the magic signature in the header.

The whole extractor was not implemented due to the script not being
maintined by Intel and any future changes in the fw_dump would mean that
the Go script (translated version of fw-dump-extractor.pl) would need to
be updated to work properly.

BUG=b:201671511, b:169152720
TEST=ran tast run ${DUT} wifi.CheckIntelFWDump

Change-Id: Ifd9985cce8d4ca00e99151bd8379cfb62a1196f8