nearbyshare: Move the CrOS setup from common to local.
chrome package is in local and we need to use that in remote tests (via
service) and in local tests. Having it in common doesn't work because it
would require a ton of refactoring. Having code that calls it in common
doesn't work because imports cannot go that way. So having the test code
that calls chrome be in local and everything else in common that is
actually directly called by remote tests works out.

BUG=b:199110070
TEST=nearbyshare.CrosToPhoneHighVis.dataoffline_allcontacts_jpg11kb

Change-Id: I547f92bc5b1921d09377283b3c3e90ba808df922