network: VPNConnect: Run l2tp_ipsec_* tests for two drivers
We are migrating the backend of L2TP/IPsec VPN from the legacy
stroke-based driver to the new swanctl-based driver. During this
period, tests need to be run for both of the two drivers. This patch can
be reverted once the migration is done.

BUG=b:204261554
TEST=tast run $DUT network.VPNConnection.l2tp_ipsec*

Cq-Depend: chromium:3256610
Change-Id: I93616a33ff564f227624fe7a54231c01f6f7a8cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3256292
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>