Tast: Increase ARCVM video encode perf test timeout.
On most platforms the current timeout of 10min is more than enough for
the ARCVM video encode performance tests. But on some lower spec boards
such as certain octopus variants this timeout might not always be
enough resulting in test flakiness. This CL increases the timeout a bit
to 12min to leave more margin for fluctuations.

BUG=b:205060753
TEST=arc.VideoEncodeAccelPerf.h264_*_i420_vm on octopus

Change-Id: Ib77bcb63046e1ffd6c155c3f161d1a034b9423b6